saving state, fully broken mid-refactor
